# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LRMC — Legacy Rental Management Consortium API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /memberships:
    get:
      tags: [membership]
      operationId: listMemberships
      summary: List memberships (admin sees all, members see their own)
      parameters:
        - name: userId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Membership"
    post:
      tags: [membership]
      operationId: createMembership
      summary: Apply for a membership
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MembershipInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Membership"
  /memberships/{id}:
    patch:
      tags: [membership]
      operationId: updateMembership
      summary: Update a membership's status (admin approval, payment confirmation)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MembershipUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Membership"

  # ---------------- Assets ----------------
...
  /land-listings:
    get:
      tags: [land]
      operationId: listLandListings
      summary: List land listings
      parameters:
        - name: sellerId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LandListing"
    post:
      tags: [land]
      operationId: createLandListing
      summary: Create a land listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LandListingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LandListing"
  /land-listings/{id}:
    patch:
      tags: [land]
      operationId: updateLandListing
      summary: Update a land listing
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LandListingUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LandListing"
  /land-transactions:
    get:
      tags: [land]
      operationId: listLandTransactions
      summary: List land transactions (brokering pipeline)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LandTransaction"
    post:
      tags: [land]
      operationId: createLandTransaction
      summary: Start a land purchase transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LandTransactionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LandTransaction"
  /land-transactions/{id}:
    patch:
      tags: [land]
      operationId: updateLandTransaction
      summary: Update a land transaction (close/cancel)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LandTransactionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LandTransaction"

  # ---------------- Construction ----------------
...
  /marketplace-listings:
    get:
      tags: [marketplace]
      operationId: listMarketplaceListings
      summary: List marketplace listings
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: sellerId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MarketplaceListing"
    post:
      tags: [marketplace]
      operationId: createMarketplaceListing
      summary: Create a marketplace listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarketplaceListingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketplaceListing"
  /marketplace-listings/{id}:
    patch:
      tags: [marketplace]
      operationId: updateMarketplaceListing
      summary: Update a marketplace listing
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarketplaceListingUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketplaceListing"
    delete:
      tags: [marketplace]
      operationId: deleteMarketplaceListing
      summary: Remove a marketplace listing
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted

  # ---------------- E-commerce (digital documents) ----------------
...
  /prospect-leads:
    get:
      tags: [prospecting]
      operationId: listProspectLeads
      summary: List prospect leads
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProspectLead"
    post:
      tags: [prospecting]
      operationId: createProspectLead
      summary: Submit a prospect lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProspectLeadInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProspectLead"
  /prospect-leads/{id}:
    patch:
      tags: [prospecting]
      operationId: updateProspectLead
      summary: Update a prospect lead's status/notes
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProspectLeadUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProspectLead"

  # ---------------- Intranet: messaging & ticketing ----------------
...
  /risk-events:
    get:
      tags: [risk]
      operationId: listRiskEvents
      summary: List risk events
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RiskEvent"
    post:
      tags: [risk]
      operationId: createRiskEvent
      summary: Log a new risk event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RiskEventInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskEvent"
  /risk-events/{id}:
    patch:
      tags: [risk]
      operationId: updateRiskEvent
      summary: Update/resolve a risk event
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RiskEventUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskEvent"

  # ---------------- Settlement ----------------
...
components:
  parameters:
...
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    Membership:
      type: object
      required: [id, userId, type, status, feeDalasi, paidAt, createdAt, updatedAt]
      properties:
        id: { type: integer }
        userId: { type: string }
        type: { type: string }
        status: { type: string }
        feeDalasi: { type: integer }
        paidAt: { type: [string, 'null'], format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    MembershipInput:
      type: object
      required: [userId, type]
      properties:
        userId: { type: string }
        type: { type: string }
        feeDalasi: { type: integer }
    MembershipUpdate:
      type: object
      properties:
        status: { type: string }
        paidAt: { type: [string, 'null'], format: date-time }

    # ---- Assets ----
...
    LandListing:
      type: object
      required: [id, sellerId, title, location, sizeAcres, priceUsd, description, imageUrl, status, createdAt, updatedAt]
      properties:
        id: { type: integer }
        sellerId: { type: string }
        title: { type: string }
        location: { type: string }
        sizeAcres: { type: integer }
        priceUsd: { type: integer }
        description: { type: string }
        imageUrl: { type: [string, 'null'] }
        status: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    LandListingInput:
      type: object
      required: [sellerId, title, location, sizeAcres, priceUsd]
      properties:
        sellerId: { type: string }
        title: { type: string, minLength: 1 }
        location: { type: string }
        sizeAcres: { type: integer }
        priceUsd: { type: integer }
        description: { type: string }
        imageUrl: { type: string }
    LandListingUpdate:
      type: object
      properties:
        title: { type: string }
        location: { type: string }
        sizeAcres: { type: integer }
        priceUsd: { type: integer }
        description: { type: string }
        imageUrl: { type: string }
        status: { type: string }
    LandTransaction:
      type: object
      required: [id, listingId, buyerId, amountUsd, status, closedAt, createdAt]
      properties:
        id: { type: integer }
        listingId: { type: integer }
        buyerId: { type: string }
        amountUsd: { type: integer }
        status: { type: string }
        closedAt: { type: [string, 'null'], format: date-time }
        createdAt: { type: string, format: date-time }
    LandTransactionInput:
      type: object
      required: [listingId, buyerId, amountUsd]
      properties:
        listingId: { type: integer }
        buyerId: { type: string }
        amountUsd: { type: integer }
    LandTransactionUpdate:
      type: object
      properties:
        status: { type: string }
        closedAt: { type: [string, 'null'], format: date-time }

    # ---- Construction ----
...
    MarketplaceListing:
      type: object
      required: [id, sellerId, title, description, category, priceDalasi, imageUrl, status, createdAt]
      properties:
        id: { type: integer }
        sellerId: { type: string }
        title: { type: string }
        description: { type: string }
        category: { type: string }
        priceDalasi: { type: integer }
        imageUrl: { type: [string, 'null'] }
        status: { type: string }
        createdAt: { type: string, format: date-time }
    MarketplaceListingInput:
      type: object
      required: [sellerId, title, priceDalasi]
      properties:
        sellerId: { type: string }
        title: { type: string, minLength: 1 }
        description: { type: string }
        category: { type: string }
        priceDalasi: { type: integer }
        imageUrl: { type: string }
    MarketplaceListingUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        category: { type: string }
        priceDalasi: { type: integer }
        imageUrl: { type: string }
        status: { type: string }

    # ---- E-commerce ----
...
    ProspectLead:
      type: object
      required: [id, name, email, phone, source, interest, status, notes, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: [string, 'null'] }
        phone: { type: [string, 'null'] }
        source: { type: string }
        interest: { type: string }
        status: { type: string }
        notes: { type: string }
        createdAt: { type: string, format: date-time }
    ProspectLeadInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        email: { type: string }
        phone: { type: string }
        source: { type: string }
        interest: { type: string }
        notes: { type: string }
    ProspectLeadUpdate:
      type: object
      properties:
        status: { type: string }
        notes: { type: string }

    # ---- Intranet ----
...
    RiskEvent:
      type: object
      required: [id, type, severity, description, status, detectedAt, resolvedAt]
      properties:
        id: { type: integer }
        type: { type: string }
        severity: { type: string }
        description: { type: string }
        status: { type: string }
        detectedAt: { type: string, format: date-time }
        resolvedAt: { type: [string, 'null'], format: date-time }
    RiskEventInput:
      type: object
      required: [type, description]
      properties:
        type: { type: string }
        severity: { type: string }
        description: { type: string, minLength: 1 }
    RiskEventUpdate:
      type: object
      properties:
        status: { type: string }
        severity: { type: string }
        resolvedAt: { type: [string, 'null'], format: date-time }

    # ---- Settlement ----
...