# lib/api-zod/src/generated/types/membershipUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * LRMC — Legacy Rental Management Consortium API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MembershipUpdate {
  status?: string;
  /** @nullable */
  paidAt?: Date | null;
}